+++ include/fkYAML/iterator.hpp
@@ -16,9 +16,9 @@
 #include <cstddef>
 #include <iterator>
 
-#include "fkYAML/VersioningMacros.hpp"
-#include "fkYAML/Exception.hpp"
-#include "fkYAML/TypeTraits.hpp"
+#include "fkYAML/version_macros.hpp"
+#include "fkYAML/exception.hpp"
+#include "fkYAML/type_traits.hpp"
 
 /**
  * @namespace fkyaml
