+++ include/fkYAML/lexical_analyzer.hpp
@@ -21,10 +21,10 @@
 #include <string>
 #include <vector>
 
-#include "fkYAML/VersioningMacros.hpp"
-#include "fkYAML/Assert.hpp"
-#include "fkYAML/Exception.hpp"
-#include "fkYAML/TypeTraits.hpp"
+#include "fkYAML/version_macros.hpp"
+#include "fkYAML/assert.hpp"
+#include "fkYAML/exception.hpp"
+#include "fkYAML/type_traits.hpp"
 
 /**
  * @namespace fkyaml
