+++ include/fkYAML/node_t.hpp
@@ -15,7 +15,7 @@
 
 #include <cstdint>
 
-#include "fkYAML/VersioningMacros.hpp"
+#include "fkYAML/version_macros.hpp"
 
 /**
  * @namespace fkyaml
